pyproject.toml
[project]
name = "armis-mcp-client"
version = "0.1.0"
description = "MCP client for Armis security platform with Ollama LLM integration"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "ollama>=0.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
